refactor: reorganize query options into respective folders for better structure
