+++ src/queries/options/bookmarks/userBookmarkedRecipesOptions.ts
@@ -1,7 +1,7 @@
 import { queryOptions } from '@tanstack/react-query';
 
 import { getUserBookmarkedRecipes } from '@/services/requests/users';
-import { queryKeys } from '../helpers';
+import { queryKeys } from '../../helpers';
 import {
   FORBIDDEN_ERROR,
   isForbiddenResponse,
